requirements.txt
# Core dependencies for demos
python-dotenv>=1.1.0
rich>=14.0.0  # For nice terminal output in verify script
anthropic>=0.5.0 # for some demos
faiss-cpu>=1.7.0 # for the FAISS demo
seaborn>=0.13.0 # for the graphing demo
fair-llm>=0.1 # fair package
pytest>=8.0.0


//forUI
fastapi>=0.112.0
uvicorn>=0.30.0
